refactor(wallet): move UnlockLockedTokens to token_lock.go
Pure code-organisation move — co-locate the real unlock operation with
the other Lock*/Unlock* functions in token_lock.go rather than leaving
it in quorum_stubs.go alongside TODO phase07 stubs.

No behaviour change; no call-site changes; unused imports removed from
quorum_stubs.go.
