+++ core/wallet/quorum_stubs.go
@@ -1,10 +1,5 @@
 package wallet
 
-import (
-	"time"
-
-	"github.com/rubixchain/rubixgoplatform/constants"
-)
 
 // AddUnpledgeSequenceInfo persists an unpledge sequence record.
 func (w *Wallet) AddUnpledgeSequenceInfo(info *UnpledgeSequenceInfo) error {
@@ -60,20 +55,6 @@ func (w *Wallet) PledgeWholeToken(did string, token string, blk interface{}) err
 func (w *Wallet) AddTokenStateHash(did string, hashes []string, pledgedTokens []string, txID string) error {
 	// TODO(phase07): insert into token_state_hashes table
 	return nil
-}
-
-// UnlockLockedTokens releases specific locked tokens for a DID back to Free status.
-// Called by the /api/unlock-tokens quorum-side endpoint when a transaction is aborted.
-func (w *Wallet) UnlockLockedTokens(did string, tokens []string) error {
-	if len(tokens) == 0 {
-		return nil
-	}
-	_, err := w.db.Pool().Exec(w.Ctx,
-		`UPDATE tokens SET token_status=$1, updated_at=$2
-		 WHERE did=$3 AND token_id = ANY($4::text[]) AND token_status=$5`,
-		constants.TokenStatus_Free, time.Now(), did, tokens, constants.TokenStatus_Locked,
-	)
-	return err
 }
 
 // RemoveTokenStateHash removes a single token state hash record.
